Problem: command warnings if no jailman user exist
even with USE_JAILER=False

Solution: This was caused by prepare_jailer() even thought it was not needed.
Made it's running conditional on use_jailer

+++ src/aleph/vm/hypervisors/firecracker/microvm.py
@@ -136,6 +136,8 @@ class MicroVM:
         }
 
     def prepare_jailer(self):
+        if not self.use_jailer:
+            return False
         system(f"rm -fr {self.jailer_path}")
 
         # system(f"rm -fr {self.jailer_path}/run/")
